+++ javascripts/navigation.js
@@ -5,18 +5,22 @@ function toggleNavigation() {
 
 function showNaviTab(tab) {
 	document.getElementById(player.navigation.naviTab).style.display = "none";
-	document.getElementById(tab).style.display = "inline-block";
-	document.getElementById(player.navigation.naviTab + "btn").className = "navitabbtn";
-	document.getElementById(tab + "btn").className = "navitabbtn active";
+	document.getElementById(tab).style.display = "";
+	document.getElementById(`${player.navigation.naviTab}btn`).classList.remove("navigation--active");
+	document.getElementById(`${tab}btn`).classList.add("navigation--active");
 	player.navigation.naviTab = tab;
 	showSubTab();
 	document.getElementById("navigation").style.display = "none";
 	document.getElementById("navi_open").style.display = "block";
 }
- function showSubTab() {
+
+function showSubTab() {
 	switch (player.navigation.naviTab) {
 		case "production_tab":
 			showProdTab(player.navigation.production);
+			break;
+		case "meltdown_tab":
+			showMeltdownTab(player.navigation.meltdown);
 			break;
 		default:
 	}
@@ -24,9 +28,9 @@ function showNaviTab(tab) {
 
 function showProdTab(tab) {
 	document.getElementById(player.navigation.production).style.display = "none";
-	document.getElementById(tab).style.display = "inline-block";
-	document.getElementById("production_" + player.navigation.production + "btn").className = "subtabbtn";
-	document.getElementById("production_" + tab + "btn").className = "subtabbtn active";
+	document.getElementById(tab).style.display = "";
+	document.getElementById(`production_${player.navigation.production}btn`).classList.remove("navigation--active");
+	document.getElementById(`production_${tab}btn`).classList.add("navigation--active");
 	player.navigation.production = tab;
 }